Practice 11

+++ Online Store Management/Services/CustomerService.cs
@@ -1,42 +1,23 @@
 ﻿using Online_Store_Management.Models;
 using Online_Store_Management.Interfaces;
 using Online_Store_Management.DataAccess;
+using Microsoft.EntityFrameworkCore;
 namespace Online_Store_Management.Services
 {
-    public class CustomerService : ICustomer
+    public class CustomerService(IRepository<CustomerDbModel> customerRepository) : ICustomer
     {
         public delegate void CustomerUpdateHandler(CustomerDbModel customerUpdate);
         public event CustomerUpdateHandler? CustomerUpdate;
-        private readonly IRepository<CustomerDbModel> customerRepository;
-        private static readonly string[] LastNamesNew = new[]
-        {
-            "Snow", "Goth", "White", "Jeffry", "Smith", "Brown"
-        };
-        private static readonly int[] PostIndexes = new[]
-        {
-            03115, 22567, 89088, 12345, 54321, 65678
-        };
-
-        private FileStream _transactionLogFileStream;
-
-        public async Task LogActionAsync(Customer customer, string message, CancellationToken cancellationToken)
-        {
-            byte[] messageBytes = System.Text.Encoding.UTF8.GetBytes($"{DateTime.UtcNow}: {message}\n");
-            _transactionLogFileStream.Write(messageBytes, 0, messageBytes.Length);
-            await Infrastructure.Logger.LogToConsole(customer, cancellationToken);
-        }
-
-        public void SetCustomerLogFileStream(FileStream customerLogFileStream)
-        {
-            this._transactionLogFileStream = customerLogFileStream;
-        }
-
-        public CustomerService(IRepository<CustomerDbModel> customerRepository)
-        {
-            this.customerRepository = customerRepository
+        private readonly IRepository<CustomerDbModel> customerRepository = customerRepository
             ?? throw new ArgumentNullException(nameof(customerRepository));
-        }
-
+        private static readonly string[] LastNamesNew = 
+        [
+            "Snow", "Goth", "White", "Jeffry", "Smith", "Brown"
+        ];
+        private static readonly int[] PostIndexes = 
+        [
+            03115, 22567, 89088, 12345, 54321, 65678
+        ];
         public async Task<CustomerDbModel?> GetCustomerByIdAsync(int id, CancellationToken cancellationToken)
         {
             return await customerRepository.GetByIdAsync(id, cancellationToken);
@@ -44,17 +25,27 @@ namespace Online_Store_Management.Services
 
         public async Task AddCustomerAsync(CustomerDbModel customer, CancellationToken cancellationToken)
         {
+            if (customer.Id < 1 || string.IsNullOrWhiteSpace(customer.LastName) || customer.PostIndex < 1)
+            {
+                throw new ArgumentException("Invalid customer data.");
+            }
+
             await customerRepository.AddAsync(customer, cancellationToken);
         }
 
         public async Task UpdateAsync(CustomerDbModel customer, CancellationToken cancellationToken)
         {
             await customerRepository.UpdateAsync(customer, cancellationToken);
-            CustomerUpdate.Invoke(customer);
+            var handler = CustomerUpdate;
+            handler?.Invoke(customer);
         }
 
         public async Task DeleteAsync(int id, CancellationToken cancellationToken)
         {
+            if (id <= 0)
+            {
+                throw new ArgumentException("Invalid ID provided.");
+            }
             await customerRepository.DeleteAsync(id, cancellationToken);
         }
 
@@ -62,7 +53,7 @@ namespace Online_Store_Management.Services
         {
             IEnumerable<string> lastNameNewQuery =
                 from name in LastNamesNew
-                where name.StartsWith("S")
+                where name.StartsWith('S')
                 select name;
 
             return lastNameNewQuery;
